Extract cloud create and delete fns
